requirements.txt
# LangChain Core
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-openai

# Environment Variable Management
python-dotenv

# Pinecone
pinecone[grpc]

# FastAPI
fastapi
uvicorn

# Utilities
requests
pypdf
streamlit